Move day view algorithm into utils and add some comments
